refactor dungen editor to own files
