+++ register_types.cpp
@@ -5,11 +5,12 @@
 #include "core/dungen.h"
 
 #ifdef TOOLS_ENABLED
-#include "editor/dungen_editor_plugin.h"
 #include "editor/components/dungen_editor_toolbar.h"
 #include "editor/components/dungen_zoom_widget.h"
 #include "editor/dungen_preview_panel.h"
 #include "editor/dungen_previewer.h"
+#include "editor/dungen_editor.h"
+#include "editor/dungen_editor_plugin.h"
 #endif // TOOLS_ENABLED
 
 #include <gdextension_interface.h>
